capstonedesign-server/requirements.txt
absl-py==2.2.1
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dlib==19.24.6
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy>=1.26.0,<2.2.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
protobuf==5.29.4
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.0
typing_extensions==4.13.0
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
# 멀티모달 감정 분석 추가 패키지
openai==1.12.0
openai-whisper==20240930
librosa==0.10.1
soundfile==0.12.1
reportlab==4.0.7
Pillow==10.2.0
pydub==0.25.1
scipy==1.12.0
scikit-learn==1.4.0
